Fix FHT to use ll

+++ code/math/fht.cpp
@@ -3,7 +3,7 @@ void fht(vi &A, bool inv=false, int l, int r) {
 	int k = (r-l)/2;
 	if (!inv) fht(A, inv, l, l+k), fht(A, inv, l+k, r);
 	rep(i,l,l+k) {
-		int x = A[i], y = A[i+k];
+		ll x = A[i], y = A[i+k];
 		if (!inv) A[i] =  x-y,    A[i+k] =   x+y;
 		else      A[i] = (x+y)/2, A[i+k] = (-x+y)/2;
 	}
